Move Logger to apis package
