+++ sources/src/main/java/com/google/solutions/jitaccess/catalog/Provisioner.java
@@ -27,6 +27,7 @@ import com.google.api.services.cloudresourcemanager.v3.model.Policy;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.util.concurrent.UncheckedExecutionException;
+import com.google.solutions.jitaccess.apis.Logger;
 import com.google.solutions.jitaccess.apis.clients.*;
 import com.google.solutions.jitaccess.catalog.auth.*;
 import com.google.solutions.jitaccess.catalog.policy.IamRoleBinding;
