+++ sources/src/main/java/com/google/solutions/jitaccess/catalog/auth/CachedSubjectResolver.java
@@ -27,7 +27,7 @@ import com.google.common.cache.LoadingCache;
 import com.google.common.util.concurrent.UncheckedExecutionException;
 import com.google.solutions.jitaccess.apis.clients.AccessException;
 import com.google.solutions.jitaccess.apis.clients.CloudIdentityGroupsClient;
-import com.google.solutions.jitaccess.catalog.Logger;
+import com.google.solutions.jitaccess.apis.Logger;
 import jakarta.inject.Singleton;
 import org.jetbrains.annotations.NotNull;
