+++ sources/src/main/java/com/google/solutions/jitaccess/catalog/auth/SubjectResolver.java
@@ -28,7 +28,7 @@ import com.google.solutions.jitaccess.apis.clients.AccessException;
 import com.google.solutions.jitaccess.apis.clients.CloudIdentityGroupsClient;
 import com.google.solutions.jitaccess.apis.clients.ResourceNotFoundException;
 import com.google.solutions.jitaccess.catalog.EventIds;
-import com.google.solutions.jitaccess.catalog.Logger;
+import com.google.solutions.jitaccess.apis.Logger;
 import com.google.solutions.jitaccess.util.CompletableFutures;
 import org.jetbrains.annotations.NotNull;
