+++ sources/src/main/java/com/google/solutions/jitaccess/catalog/legacy/LegacyPolicyLoader.java
@@ -27,7 +27,7 @@ import com.google.solutions.jitaccess.apis.clients.AccessException;
 import com.google.solutions.jitaccess.apis.clients.AssetInventoryClient;
 import com.google.solutions.jitaccess.apis.clients.ResourceManagerClient;
 import com.google.solutions.jitaccess.apis.clients.ResourceNotFoundException;
-import com.google.solutions.jitaccess.catalog.Logger;
+import com.google.solutions.jitaccess.apis.Logger;
 import com.google.solutions.jitaccess.catalog.policy.Policy;
 import org.jetbrains.annotations.NotNull;
