+++ sources/src/main/java/com/google/solutions/jitaccess/web/rest/HealthResource.java
@@ -22,7 +22,7 @@
 package com.google.solutions.jitaccess.web.rest;
 
 import com.google.solutions.jitaccess.apis.clients.Diagnosable;
-import com.google.solutions.jitaccess.catalog.Logger;
+import com.google.solutions.jitaccess.apis.Logger;
 import com.google.solutions.jitaccess.util.CompletableFutures;
 import com.google.solutions.jitaccess.web.EventIds;
 import jakarta.enterprise.context.Dependent;
