+++ sources/src/test/java/com/google/solutions/jitaccess/web/TestStructuredLogger.java
@@ -21,14 +21,8 @@
 
 package com.google.solutions.jitaccess.web;
 
-import com.google.solutions.jitaccess.catalog.auth.Directory;
-import com.google.solutions.jitaccess.catalog.auth.SubjectResolver;
-import com.google.solutions.jitaccess.catalog.auth.EndUserId;
-import org.junit.jupiter.api.Nested;
+import com.google.solutions.jitaccess.apis.StructuredLogger;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -44,7 +38,8 @@ public class TestStructuredLogger {
     logger.info("E1", "message");
 
     assertEquals(
-      "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\",\"event/type\":\"operational\"},\"message\":\"message\"}\n",
+      "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\"," +
+        "\"event/type\":\"operational\"},\"message\":\"message\"}\n",
       buffer.toString());
   }
 
@@ -55,7 +50,8 @@ public class TestStructuredLogger {
     logger.info("E1", "s=%s, d=%d", "test", 1);
 
     assertEquals(
-      "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\",\"event/type\":\"operational\"},\"message\":\"s=test, d=1\"}\n",
+      "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\"," +
+        "\"event/type\":\"operational\"},\"message\":\"s=test, d=1\"}\n",
       buffer.toString());
   }
 
@@ -70,7 +66,8 @@ public class TestStructuredLogger {
     logger.warn("E1", "message");
 
     assertEquals(
-      "{\"severity\":\"WARN\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\",\"event/type\":\"operational\"},\"message\":\"message\"}\n",
+      "{\"severity\":\"WARN\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\"," +
+        "\"event/type\":\"operational\"},\"message\":\"message\"}\n",
       buffer.toString());
   }
 
@@ -81,7 +78,8 @@ public class TestStructuredLogger {
     logger.warn("E1", "s=%s, d=%d", "test", 1);
 
     assertEquals(
-      "{\"severity\":\"WARN\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\",\"event/type\":\"operational\"},\"message\":\"s=test, d=1\"}\n",
+      "{\"severity\":\"WARN\",\"logging.googleapis.com/labels\":{\"event/id\":\"E1\"," +
+        "\"event/type\":\"operational\"},\"message\":\"s=test, d=1\"}\n",
       buffer.toString());
   }
 
@@ -136,74 +134,5 @@ public class TestStructuredLogger {
 
     assertTrue(buffer.toString()
       .contains("exception: outer-exception, caused by IllegalArgumentException: inner-exception"));
-  }
-
-
-  // -------------------------------------------------------------------------
-  // Inner classes.
-  // -------------------------------------------------------------------------
-
-  @Nested
-  public static class RequestContextLogger {
-
-    @Test
-    public void info_whenTraceIdAndUserIdSet() {
-      var buffer = new StringBuilder();
-      var requestContext = new RequestContext(Mockito.mock(SubjectResolver.class));
-      requestContext.initialize("GET", "/", "trace-1");
-      requestContext.authenticate(
-        new EndUserId("id"),
-        Directory.CONSUMER,
-        new IapDevice("device-id", List.of()));
-      var logger = new StructuredLogger.RequestContextLogger(buffer, requestContext);
-
-      logger.info("event-1", "message-1");
-
-      assertEquals(
-        "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":" +
-          "{\"auth/device_id\":\"device-id\",\"event/id\":\"event-1\"," +
-          "\"auth/user_id\":\"id\",\"event/type\":\"operational\"," +
-          "\"auth/directory\":\"CONSUMER\"," +
-          "\"auth/device_access_levels\":\"\",\"request/path\":\"/\"," +
-          "\"request/method\":\"GET\"},\"message\":\"message-1\",\"logging.googleapis.com/trace\":\"trace-1\"}\n",
-        buffer.toString());
-    }
-
-    @Test
-    public void info_whenTraceIdAndAccessLevelsSet() {
-      var buffer = new StringBuilder();
-      var requestContext = new RequestContext(Mockito.mock(SubjectResolver.class));
-      requestContext.initialize("GET", "/", "trace-1");
-      requestContext.authenticate(
-        new EndUserId("id"),
-        Directory.CONSUMER,
-        new IapDevice("device-id", List.of("level-1", "level-2")));
-      var logger = new StructuredLogger.RequestContextLogger(buffer, requestContext);
-
-      logger.info("event-1", "message-1");
-
-      assertEquals(
-        "{\"severity\":\"INFO\",\"logging.googleapis.com/labels\":" +
-          "{\"auth/device_id\":\"device-id\",\"event/id\":\"event-1\"," +
-          "\"auth/user_id\":\"id\",\"event/type\":\"operational\"," +
-          "\"auth/directory\":\"CONSUMER\"," +
-          "\"auth/device_access_levels\":\"level-1, level-2\"," +
-          "\"request/path\":\"/\",\"request/method\":\"GET\"}," +
-          "\"message\":\"message-1\",\"logging.googleapis.com/trace\":\"trace-1\"}\n",
-        buffer.toString());
-    }
-
-    @Test
-    public void error_whenNotAuthenticated() {
-      var buffer = new StringBuilder();
-      var requestContext = new RequestContext(Mockito.mock(SubjectResolver.class));
-      var logger = new StructuredLogger.RequestContextLogger(buffer, requestContext);
-      logger.error("event-1", "message-1");
-
-      assertEquals(
-        "{\"severity\":\"ERROR\",\"logging.googleapis.com/labels\":{\"event/id\":\"event-1\",\"event/type\":\"operational\"}," +
-          "\"message\":\"message-1\"}\n",
-        buffer.toString());
-    }
   }
 }
