+++ sources/src/test/java/com/google/solutions/jitaccess/web/rest/TestOperationAuditTrail.java
@@ -22,7 +22,7 @@
 package com.google.solutions.jitaccess.web.rest;
 
 import com.google.solutions.jitaccess.catalog.JitGroupContext;
-import com.google.solutions.jitaccess.catalog.Logger;
+import com.google.solutions.jitaccess.apis.Logger;
 import com.google.solutions.jitaccess.catalog.auth.EndUserId;
 import com.google.solutions.jitaccess.catalog.auth.JitGroupId;
 import com.google.solutions.jitaccess.catalog.auth.Principal;
